User, seller

+++ src/main/java/com/codeblue/montreISTA/service/implement/SellerServiceImpl.java
@@ -1,13 +1,16 @@
 package com.codeblue.montreISTA.service.implement;
 
+import com.codeblue.montreISTA.DTO.LoginSellerResponseDTO;
 import com.codeblue.montreISTA.DTO.SellerRequestDTO;
 import com.codeblue.montreISTA.DTO.SellerResponseDTO;
 import com.codeblue.montreISTA.entity.Product;
+import com.codeblue.montreISTA.entity.Role;
 import com.codeblue.montreISTA.entity.Seller;
 import com.codeblue.montreISTA.entity.User;
 import com.codeblue.montreISTA.helper.DTOConverter;
 import com.codeblue.montreISTA.helper.Pagination;
 import com.codeblue.montreISTA.repository.ProductRepository;
+import com.codeblue.montreISTA.repository.RoleRepository;
 import com.codeblue.montreISTA.repository.SellerRepository;
 import com.codeblue.montreISTA.repository.UserRepository;
 import com.codeblue.montreISTA.service.SellerService;
@@ -29,6 +32,7 @@ public class SellerServiceImpl implements SellerService {
     private final SellerRepository sellerRepository;
     private final ProductRepository productRepository;
     private final CloudinaryService cloudinaryService;
+    private final RoleRepository roleRepository;
 
 
     @Override
@@ -44,27 +48,20 @@ public class SellerServiceImpl implements SellerService {
     }
 
     @Override
-    public Object createSeller(SellerRequestDTO sellerRequestDTO, Authentication authentication) throws Exception  {
+    public SellerResponseDTO createSeller(SellerRequestDTO sellerRequestDTO, Authentication authentication) throws Exception  {
         User user = userRepository.findByUsername(authentication.getName()).orElseThrow(()->new Exception("Please sign up"));
         Optional<Seller> sellerOptional = sellerRepository.findByUserUserId(user.getUserId());
         Optional<Seller> sellerValidation = sellerRepository.findByStoreName(sellerRequestDTO.getStoreName());
+        if(sellerOptional.isPresent()) {
+            throw new Exception("You only can have 1 store");
+        }
         if(sellerValidation.isPresent()){
             throw new Exception("Your Store Name has been used");
         }
-        if(authentication==null){
-            throw new Exception("Please login");
-        }
-
         Seller seller = sellerRequestDTO.convertToEntity(user);
         seller.setStorePhoto("https://www.shutterstock.com/image-vector/shop-icon-store-230619400");
-        if(sellerOptional.isPresent()){
-            Pageable pageable = Pagination.paginate(0, "price", false);
-            List<Product> products = productRepository.findBySellerSellerId(sellerOptional.get().getSellerId(), pageable);
-            return DTOConverter.convertProducts(products);
-        }else {
-            Seller sellerDTO = sellerRepository.save(seller);
-            return sellerDTO.convertToResponse();
-        }
+        Seller sellerDTO = sellerRepository.save(seller);
+        return sellerDTO.convertToResponse();
     }
     @Override
     public SellerResponseDTO updateSeller(SellerRequestDTO seller, Authentication authentication)throws Exception {
@@ -84,14 +81,30 @@ public class SellerServiceImpl implements SellerService {
     }
 
     @Override
-    public void deleteSeller(Long id) {
-        sellerRepository.deleteById(id);
+    public LoginSellerResponseDTO loginAsSeller(String keyword,Integer page, String sort, boolean descending) throws Exception {
+        Pageable pageable = Pagination.paginate(page, sort, descending);
+        Seller seller = sellerRepository.findByUserUsername(keyword).orElseThrow(()->new Exception("Create your shop first"));
+        List<Product> products = productRepository.findBySellerUserNameIgnoreCaseContaining(keyword,pageable);
+        return DTOConverter.convertLoginSeller(seller,products);
+    }
+
+    @Override
+    public void deleteSeller(Long id,Authentication authentication) throws Exception{
+        Seller seller = sellerRepository.findById(id).orElseThrow(()->new Exception("Seller not found"));
+        List<Role> roles = roleRepository.findByUsersUserUsername(authentication.getName());
+        boolean checkRoles = roles.stream().anyMatch(role -> role.getRoleName().equals("ROLE_ADMIN"));
+        boolean checkUser = seller.getUser().getUsername().equals(authentication.getName());
+        if (checkRoles || checkUser){
+            sellerRepository.deleteById(id);
+        } else {
+            throw new Exception("You can't delete other seller");
+        }
     }
 
 
     @Override
     public SellerResponseDTO findByUsername(String keyword)throws Exception {
-        Seller seller = sellerRepository.findByUserUsername(keyword).orElseThrow(()->new Exception("Seller not found"));
+        Seller seller = sellerRepository.findByUserUsername(keyword).orElseThrow(()->new Exception("Create your shop first"));
         return seller.convertToResponse();
     }
